+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- fixes "Information" category mapping if no "Support" category is present (#36)
 
 ## [v1.5.0](https://github.com/cloudogu/warp-menu/releases/tag/v1.5.0)
 ### Added